chore: Update bentoo.sh

+++ scripts/bentoo.sh
@@ -6,6 +6,15 @@ ARCH=$(uname -m)
 
 cd "$BASE_DIR" || exit;
 
+UPDATE_BENTOO="false"
+
+for OPT in "$@"; do
+  if [[ "$OPT" = *"--update"* ]]; then
+      UPDATE_BENTOO="true"
+      shift 1
+  fi
+done
+
 if [[ "$OSTYPE" == "darwin"* ]]; then
     if [[ "$ARCH" == "x86_64" ]]; then
         echo "Downloading bentoo binary for Mac OS / x86_64"
@@ -34,7 +43,9 @@ wget https://github.com/flawgarden/bentoo/releases/download/latest/$ARCHIVE_NAME
 tar -xf $ARCHIVE_NAME
 rm $ARCHIVE_NAME
 
-# Download tool runners
-wget https://github.com/flawgarden/bentoo/releases/download/latest/tool_runners.tar.gz
-tar -xf tool_runners.tar.gz
-rm tool_runners.tar.gz
+if [[ "$UPDATE_BENCHMARKS" = "true" ]]; then
+    # Download tool runners
+    wget https://github.com/flawgarden/bentoo/releases/download/latest/tool_runners.tar.gz
+    tar -xf tool_runners.tar.gz
+    rm tool_runners.tar.gz
+fi
